+++ application/controllers/StatisticsCliController.php
@@ -264,5 +264,52 @@ class StatisticsCliController extends IXP_Controller_CliAction
     }
     
     
+    public function uploadTrafficStatsToDbAction()
+    {
+        // This should only be done once a day and if values already exist for 'today',
+        // just delete them.
+        $day = date( 'Y-m-d' );
+        $this->getD2EM()->getRepository( '\\Entities\\TrafficDaily' )->deleteForDay( $day );
+    
+        $custs = $this->getD2EM()->getRepository( '\\Entities\\Customer' )->getCurrentActive( false, true, true );
+    
+        foreach( $custs as $cust )
+        {
+            $stats = array();
+    
+            foreach( IXP_Mrtg::$CATEGORIES as $category )
+            {
+                $mrtg = new IXP_Mrtg(
+                        IXP_Mrtg::getMrtgFilePath( $this->_options['mrtg']['path'] . '/members',
+                                'LOG', 'aggregate', $category,
+                                $cust->getShortname()
+                        )
+                );
+    
+                $td = new \Entities\TrafficDaily();
+                $td->setDay( new DateTime( $day ) );
+                $td->setCategory( $category );
+                $td->setCustomer( $cust );
+    
+                foreach( IXP_Mrtg::$PERIODS as $name => $period )
+                {
+                    $stats = $mrtg->getValues( $period, $category, false );
+    
+                    $fn = "set{$name}AvgIn";  $td->$fn( $stats['averagein']  );
+                    $fn = "set{$name}AvgOut"; $td->$fn( $stats['averageout'] );
+                    $fn = "set{$name}MaxIn";  $td->$fn( $stats['maxin']      );
+                    $fn = "set{$name}MaxOut"; $td->$fn( $stats['maxout']     );
+                    $fn = "set{$name}TotIn";  $td->$fn( $stats['totalin']    );
+                    $fn = "set{$name}TotOut"; $td->$fn( $stats['totalout']   );
+                }
+    
+                $this->getD2EM()->persist( $td );
+            }
+            $this->getD2EM()->flush();
+        }
+    }
+    
+    
+    
 }